♻️ combine builds for emulated wheels (#22)
Instead of `3 * |supported python versions|` jobs, this will only create `3` jobs to build wheels on platforms that need emulation.
Tools running this workflow and are using scikit-build-core as a build backend, should set

```toml
[tool.scikit-build]
# ...
build-dir = "build/{build_type}"
# ...
```

so that the build directory can be reused across builds. This only requires the actual binding target to be recompiled.

It is recommended to also set
```toml
[tool.cibuildwheel.windows]
# ...
environment = { SKBUILD_CMAKE_ARGS="--fresh" }
# ...
```

to avoid problems on Windows.